# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RS Mas Amsyar WEB API",
        "description": "\r\n * > Footprint 1.0 (2024 June 25) `Developed along with the web frontend to be consumed by and also enabled integration with 3rd party API in the future`\r\n * > Footprint x (2024 xx xx) `this-is-footnote-format-please-continue-on-next-dev`",
        "contact": {
            "name": "Evelline Krist.",
            "email": "ev.attoff@gmail.com"
        },
        "version": "VER 1.0",
        "x-logo": {
            "url": "https://upload.wikimedia.org/wikipedia/commons/thumb/9/9a/Laravel.svg/985px-Laravel.svg.png"
        }
    },
    "paths": {
        "/api/article": {
            "get": {
                "tags": [
                    "Article",
                    "News"
                ],
                "summary": "Display a listing of items",
                "operationId": "articleIndex",
                "parameters": [
                    {
                        "name": "_page",
                        "in": "query",
                        "description": "current page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    },
                    {
                        "name": "_limit",
                        "in": "query",
                        "description": "max item in a page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 10
                        }
                    },
                    {
                        "name": "_search",
                        "in": "query",
                        "description": "word to search",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "_dir",
                        "in": "query",
                        "description": "order by direction",
                        "required": false,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Article",
                    "News"
                ],
                "summary": "Store a newly created item",
                "operationId": "articleStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "type",
                                    "title"
                                ],
                                "properties": {
                                    "portal_user_id": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "sequence": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "news"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "submited"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "title-of-the-news"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Title of The News"
                                    },
                                    "keywords": {
                                        "type": "string",
                                        "example": "keyword,keyword2,keyword3"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "<p><i>`Neque porro quisquam est qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit...`</i></p>"
                                    },
                                    "img_main": {
                                        "description": "Image to upload",
                                        "type": "file"
                                    },
                                    "caption": {
                                        "type": "string",
                                        "example": "here is the caption..."
                                    },
                                    "post_at": {
                                        "type": "string",
                                        "example": "2024-06-28 08:00:00"
                                    },
                                    "publish_at": {
                                        "type": "string",
                                        "example": "2024-06-28 08:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            }
        },
        "/api/article/{id}": {
            "get": {
                "tags": [
                    "Article",
                    "News"
                ],
                "summary": "Display the specified item",
                "operationId": "articleShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of item that needs to be displayed",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Article",
                    "News"
                ],
                "summary": "Update the specified item",
                "operationId": "articleUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of item that needs to be updated",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "clean-eating"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "type",
                                    "title"
                                ],
                                "properties": {
                                    "portal_user_id": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "sequence": {
                                        "type": "integer",
                                        "example": ""
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "news"
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "submited"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "title-of-the-news-edited"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "[EDITED] Title of The News"
                                    },
                                    "keywords": {
                                        "type": "string",
                                        "example": "keyword,keyword2,keyword3,some-new-keyword-because-now-its-edited"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "<p>[EDITED]</p><p><i>`Neque porro quisquam est qui dolorem ipsum quia dolor sit amet, consectetur, adipisci velit...`</i></p>"
                                    },
                                    "img_main": {
                                        "description": "Image to upload",
                                        "type": "file"
                                    },
                                    "caption": {
                                        "type": "string",
                                        "example": "[EDITED] here is the caption..."
                                    },
                                    "post_at": {
                                        "type": "string",
                                        "example": "2024-06-28 08:00:00"
                                    },
                                    "publish_at": {
                                        "type": "string",
                                        "example": "2024-06-28 08:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Article",
                    "News"
                ],
                "summary": "Remove the specified item",
                "operationId": "articleDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifier of item that needs to be removed",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Item not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport_token_ready": [],
                        "passport": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "passport_token_ready": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format without \"Bearer \"",
                "name": "Authorization",
                "in": "header"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}